server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
email-validator==2.1.0
pydantic==2.5.0
aiofiles==23.2.1
PyJWT==2.8.0